rename files to match components better
